backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
